+++ Sources/Models/DeviceStatus.swift
@@ -1,0 +1,13 @@
+//
+//  DeviceStatus.swift
+//  AppStoreConnect-Swift-SDK
+//
+//  Created by Patrick Balestra on 12/24/19.
+//
+
+import Foundation
+
+public enum DeviceStatus: String, Codable {
+    case enabled = "ENABLED"
+    case disabled = "DISABLED"
+}