+++ Changelog.md
@@ -5,6 +5,8 @@
 - Add support for all [Certificates](https://developer.apple.com/documentation/appstoreconnectapi/certificates) endpoints. (@BalestraPatrick)
 - Add support for all [Bundle ID](https://developer.apple.com/documentation/appstoreconnectapi/bundle_ids) endpoints. (@ruipfcosta, @BalestraPatrick)
 - Add support for all [Bundle ID Capabilities](https://developer.apple.com/documentation/appstoreconnectapi/bundle_id_capabilities) endpoints. (@BalestraPatrick)
+- Add support for all [Profiles](https://developer.apple.com/documentation/appstoreconnectapi/profiles) endpoints. (@@BalestraPatrick)
+- Add support for all [Devices](https://developer.apple.com/documentation/appstoreconnectapi/devices) endpoints. (@BalestraPatrick)
 - Updated CI to make sure PR requests are reviewed using Danger. (@AvdLee)
 - Removed custom Result type. (@ruipfcosta)
 - Fixed typo for downloadSalesAndTrendsReports ([Issue 42](https://github.com/AvdLee/appstoreconnect-swift-sdk/issues/42))